Rewritten the way the players tools are managed.
Now, each tool is a class in the package Tools, just like the classes. These classes have to extends PlayerTool.
A tool have an ID, used internally to identofy it. These tools needs to be registered into the ToolsManager, and into an enum listing all tools' IDs.

I'm not very happy with this way to get the tools (enum + registration in the class + separated ID): we need a way to identify the tools and the name cannot be used (formatting, this stuff); this said, @Infinity, if you have a better idea implement it!

Also, please check all the TODOs and other FIXMEs comments, especially the one concerning the CountdownTimer (TODO: switch from slot to PlayerTool with the slot cached somewhere).

*This awkward moment when almost every single file of the project is modified. #thisMoment*